requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
celery
redis
drf-yasg==1.21.7

